[GOHW-4]: pass valid order to storage; add checks to gateway

+++ internal/api/gateway/server/delete.go
@@ -2,19 +2,20 @@ package server
 
 import (
 	"context"
-	"github.com/pkg/errors"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/types/known/emptypb"
-	"homework/internal/storage/connections"
 	pb "homework/pkg/api"
 )
 
 func (i *gatewayServer) MovieDelete(ctx context.Context, req *pb.MovieDeleteRequest) (*emptypb.Empty, error) {
+	id := req.GetId()
+	if id <= 0 {
+		return nil, status.Error(codes.InvalidArgument, "id must be > 0")
+	}
+
 	if err := i.storage.Delete(ctx, req.GetId()); err != nil {
-		return nil, status.Error(codes.Internal, err.Error())
-	} else if errors.Is(err, connections.ErrTimeout) {
-		return nil, status.Error(codes.DeadlineExceeded, err.Error())
+		return nil, err
 	}
 
 	return &emptypb.Empty{}, nil
